refactor: jsx 파일을 tsx로 변경
